refactor: move getAnalysisCompletion into service
Created new service module for CdGpt and moved entryAnalysis.getAnalysisCompletion
into it. Refactored unit tests and dependencies on entryAnalysis.getAnalysisCompletion
